Fixing issues per Enno suggestions
