+++ avroparquet/src/main/scala/akka/stream/alpakka/avroparquet/javadsl/AvroParquetSink.scala
@@ -5,9 +5,10 @@
 package akka.stream.alpakka.avroparquet.javadsl
 
 import java.util.concurrent.CompletionStage
-import akka.stream.alpakka.avroparquet.scaladsl.AvroParquetFlow
-import akka.{Done, NotUsed}
+
+import akka.stream.alpakka.avroparquet.impl.AvroParquetFlow
 import akka.stream.javadsl.{Flow, Keep, Sink}
+import akka.{Done, NotUsed}
 import org.apache.avro.generic.GenericRecord
 import org.apache.parquet.hadoop.ParquetWriter
