+++ avroparquet/src/main/scala/akka/stream/alpakka/avroparquet/javadsl/AvroParquetSource.scala
@@ -2,45 +2,17 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.avroparquet.scaladsl
+package akka.stream.alpakka.avroparquet.javadsl
 
 import akka.NotUsed
-import akka.stream.{Attributes, Outlet, SourceShape}
-import akka.stream.scaladsl.Source
-import akka.stream.stage.{GraphStage, GraphStageLogic, OutHandler}
+import akka.stream.javadsl.Source
 import org.apache.avro.generic.GenericRecord
 import org.apache.parquet.hadoop.ParquetReader
 
 object AvroParquetSource {
 
-  def apply(reader: ParquetReader[GenericRecord]): Source[GenericRecord, NotUsed] =
-    Source.fromGraph(new AvroParquetSource(reader))
+  def create(reader: ParquetReader[GenericRecord]): Source[GenericRecord, NotUsed] =
+    Source.fromGraph(new akka.stream.alpakka.avroparquet.impl.AvroParquetSource(reader))
 }
 
-class AvroParquetSource(reader: ParquetReader[GenericRecord]) extends GraphStage[SourceShape[GenericRecord]] {
 
-  val out: Outlet[GenericRecord] = Outlet("AvroParquetSource")
-
-  override def createLogic(inheritedAttributes: Attributes): GraphStageLogic = new GraphStageLogic(shape) {
-
-    setHandler(
-      out,
-      new OutHandler {
-
-        override def onDownstreamFinish(): Unit = {
-          super.onDownstreamFinish()
-          reader.close()
-        }
-
-        override def onPull(): Unit = {
-          val record = reader.read()
-          Option(record).fold {
-            reader.close()
-            complete(out)
-          }(push(out, _))
-        }
-      }
-    )
-  }
-  override def shape: SourceShape[GenericRecord] = SourceShape.of(out)
-}
