+++ avroparquet/src/main/scala/akka/stream/alpakka/avroparquet/scaladsl/AvroParquetSource.scala
@@ -2,15 +2,15 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.avroparquet.javadsl
+package akka.stream.alpakka.avroparquet.scaladsl
 import akka.NotUsed
-import akka.stream.javadsl.Source
+import akka.stream.scaladsl.Source
 import org.apache.avro.generic.GenericRecord
 import org.apache.parquet.hadoop.ParquetReader
 
 object AvroParquetSource {
 
-  def create(reader: ParquetReader[GenericRecord]): Source[GenericRecord, NotUsed] =
-    Source.fromGraph(new akka.stream.alpakka.avroparquet.scaladsl.AvroParquetSource(reader))
+  def apply(reader: ParquetReader[GenericRecord]): Source[GenericRecord, NotUsed] =
+    Source.fromGraph(new akka.stream.alpakka.avroparquet.impl.AvroParquetSource(reader))
 
 }
