+++ avroparquet/src/test/java/akka/stream/alpakka/avroparquet/javadsl/AvroParquetSinkTest.java
@@ -5,7 +5,6 @@
 package akka.stream.alpakka.avroparquet.javadsl;
 
 import akka.Done;
-import akka.NotUsed;
 import akka.actor.ActorSystem;
 import akka.stream.ActorMaterializer;
 import akka.stream.Materializer;
